% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esios-archive.R
\name{esios_archives}
\alias{esios_archives}
\title{Retrieve archives from ESIOS}
\usage{
esios_archives()
}
\value{
A data.frame with the information of what is available and where.
}
\description{
Check which archives are available.
}
\examples{
\donttest{
archives <- esios_archives()
head(archives)
}
}
