% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esios-indicators.R
\name{esios_indicators}
\alias{esios_indicators}
\title{Retrieve an indicator}
\usage{
esios_indicators(
  indicator,
  locale = NULL,
  datetime = NULL,
  start_date = NULL,
  end_date = NULL,
  time_agg = NULL,
  time_trunc = NULL,
  geo_agg = NULL,
  geo_ids = NULL,
  geo_trunc = NULL
)
}
\arguments{
\item{indicator}{ID code for an indicator}

\item{locale}{Get translations for sources (es, en). Default language: es}

\item{datetime}{A certain date to filter values by (iso8601 format)}

\item{start_date}{Beginning of the date range to filter indicator values
(iso8601 format)}

\item{end_date}{End of the date range to filter indicator values
(iso8601 format)}

\item{time_agg}{How to aggregate indicator values when grouping them by time.
Accepted values: 'sum', 'average'. Default value: 'sum'.}

\item{time_trunc}{Tells the API how to trunc data time series. Accepted
values: 'five_minutes', 'ten_minutes', 'fifteen_minutes', 'hour', 'day', 'month', 'year'.}

\item{geo_agg}{How to aggregate indicator values when grouping them by geo_id.
Accepted values: 'sum', 'average'. Default value: 'sum'.}

\item{geo_ids}{Tells the API the geo ids to filter the date.}

\item{geo_trunc}{Tells the API how to group data at geolocalization level
when the geo_agg is informed. Accepted values: 'country', 'electric_system',
'autonomous_community', 'province', 'electric_subsystem', 'town' and
'drainage_basin'.}
}
\description{
Retrieve the information of any indicator based on the id.
}
\examples{
\donttest{
ei10001 <- esios_indicators(10001)
ei1700 <- esios_indicators(1700)
}
}
\references{
\url{https://api.esios.ree.es/}
}
\seealso{
\code{\link[=esios_search_indicators]{esios_search_indicators()}}
}
