% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esios-indicators.R
\name{esios_pvp}
\alias{esios_pvp}
\title{Prediction of electricity prices using indicator 1001}
\usage{
esios_pvp(
  locale = NULL,
  datetime = NULL,
  start_date = NULL,
  end_date = NULL,
  time_agg = NULL,
  time_trunc = NULL,
  geo_agg = NULL,
  geo_ids = NULL,
  geo_trunc = NULL
)
}
\arguments{
\item{locale}{Get translations for sources (es, en). Default language: es}

\item{datetime}{A certain date to filter values by (iso8601 format)}

\item{start_date}{Beginning of the date range to filter indicator values
(iso8601 format)}

\item{end_date}{End of the date range to filter indicator values
(iso8601 format)}

\item{time_agg}{How to aggregate indicator values when grouping them by time.
Accepted values: 'sum', 'average'. Default value: 'sum'.}

\item{time_trunc}{Tells the API how to trunc data time series. Accepted
values: 'five_minutes', 'ten_minutes', 'fifteen_minutes', 'hour', 'day', 'month', 'year'.}

\item{geo_agg}{How to aggregate indicator values when grouping them by geo_id.
Accepted values: 'sum', 'average'. Default value: 'sum'.}

\item{geo_ids}{Tells the API the geo ids to filter the date.}

\item{geo_trunc}{Tells the API how to group data at geolocalization level
when the geo_agg is informed. Accepted values: 'country', 'electric_system',
'autonomous_community', 'province', 'electric_subsystem', 'town' and
'drainage_basin'.}
}
\value{
A data.frame with value ( €/MWh), datetime, datetime_utc, tz_time,
geo_id and geo_name.
}
\description{
This is a shortcut for \code{esios_indicators("1001")} with some parsing.
}
\examples{
e <- esios_pvp()
}
\seealso{
\code{\link[=esios_indicators]{esios_indicators()}}
}
