% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esios-indicators.R
\name{esios_search_indicators}
\alias{esios_search_indicators}
\title{List ESIOS indicators}
\usage{
esios_search_indicators(
  text = NULL,
  taxonomy_terms = NULL,
  taxonomy_ids = NULL
)
}
\arguments{
\item{text}{Text to search indicators}

\item{taxonomy_terms}{Terms of indicators}

\item{taxonomy_ids}{Ids of the indicators}
}
\value{
A data.frame with four columns: name, description, short_name and id.
}
\description{
Find which indicators are available.
}
\examples{
\donttest{
ei <- esios_search_indicators()
mercados <- esios_search_indicators(text = "Mercados y precios")
}
}
\seealso{
\code{\link[=esios_indicators]{esios_indicators()}}
}
