% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ree-apidatos.R
\name{get_ree_ccaa}
\alias{get_ree_ccaa}
\title{Get all the CCAA available}
\usage{
get_ree_ccaa()
}
\value{
A vector with the regions
}
\description{
Get all the CCAA available
}
\examples{
get_ree_ccaa()
}
\seealso{
\code{\link[=ree_call]{ree_call()}}
}
