% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_token.R
\name{get_token}
\alias{get_token}
\title{Get ESIOS token}
\usage{
get_token(token = NULL)
}
\arguments{
\item{token}{The character string as a token}
}
\value{
Invisible the token
}
\description{
To get a token you need to ask for it.
}
\examples{
token <- get_token("aahdahgagdadfsafd")
}
