% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ree-apidatos.R
\name{ree_call}
\alias{ree_call}
\title{Retrieve data from Red Electrica Española.}
\usage{
ree_call(
  category,
  widget,
  start_date = end_date - 1,
  end_date = Sys.Date(),
  time_trunc = NULL,
  geo_trunc = NULL,
  region = "peninsular",
  lang = "es",
  parse = TRUE,
  ...
)
}
\arguments{
\item{category}{Type of data, one of c("balance", "demanda", "generacion", "intercambios",
"transporte").}

\item{widget}{Defines what to retrieve.}

\item{start_date}{Date}

\item{end_date}{Date}

\item{time_trunc}{One of \code{c("hour", "day", "month", "year")}}

\item{geo_trunc}{Only accepts "electric_system".}

\item{region}{One of \code{c("peninsular", "canarias", "baleares", "ceuta", "melilla")} or the official name of a CCAA (one of \code{get_ree_ccaa()}).}

\item{lang}{Language: either "es" or "en".}

\item{parse}{Decide if the function should try to parse the result or not
(By default TRUE).}

\item{...}{Other arguments passed to \code{httr2::req_url_query()}.}
}
\value{
A data.frame by default or a list if \code{parse = FALSE}.
}
\description{
Retrieve data from Red Electrica Española.
}
\examples{
rc <- ree_call(start_date = as.Date("2018-01-01"),
               end_date = "2018-12-31T23:59",
               time_trunc = "month",
               region = "peninsular", lang = "en",
               category = "demanda", widget = "ire-general")
rc2 <- ree_call(category = "balance", widget = "balance-electrico",
                start_date = "2022-12-31T23:59",
                region = "peninsular", lang = "en", time_trunc = "day")
rc3 <- ree_call(lang = "es",
                category = "generacion",
                widget = "estructura-generacion",
                start_date = "2014-01-01T00:00",
                end_date = "2018-12-31T23:59",
                time_trunc = "year",
                geo_trunc = "electric_system",
                region = "Castilla la Mancha")
}
