% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resios-package.R
\docType{package}
\name{resios-package}
\alias{resios}
\alias{resios-package}
\title{Accessing the electricity data of official organizations of Spain.}
\description{
Provides functions to retrieve and use easily data from Red Electrica Española (REE) the which operates the national electricity grid in Spain. The REE also provides ESIOS from electronic-"Sistema de Información del Operador del Sistema", which is also covered here.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenspain/resios}
  \item \url{https://ropenspain.github.io/resios/}
  \item Report bugs at \url{https://github.com/ropenspain/resios/issues}
}

}
\author{
\strong{Maintainer}: Lluís Revilla Sancho \email{lluis.revilla@gmail.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})

}
\keyword{internal}
